% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmpkg.geouk.R
\docType{package}
\name{dmpkg.geouk}
\alias{dmpkg.geouk}
\title{dmpkg.geouk: This package contains the digital vector boundaries, in `sf`format and "simplified" at 20%, for main British geographies, (BGC) Generalised (20m), clipped to the coastline.}
