% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entities}
\alias{entities}
\title{Entities}
\format{
A data.table with the following ??? column:
\describe{
  \item{\code{location_type}}{An acronym for the geography, used as its id}
  \item{\code{country}}{}
  \item{\code{prog_id}}{}
  \item{\code{ordering}}{}
  \item{\code{code}}{}
  \item{\code{name}}{}
  \item{\code{ons}}{}
  \item{\code{N_ons}}{}
  \item{\code{N_onspd}}{}
  \item{\code{N_package}}{}
}

For further details, see \url{http://}
}
\usage{
entities
}
\description{
An alternative view of the geographies included in the package, as listed in \code{location_types}, partitioned by countries
}
\keyword{datasets}
