% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_types}
\alias{location_types}
\title{Location Types}
\format{
A data.table with the following ??? columns:
\describe{
  \item{\code{location_type}}{An acronym for the geography, used as its id}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
}

For further details, see \url{http://}
}
\usage{
location_types
}
\description{
A list of all the Geographies included in the package
(apart from the Output Areas and the Workplace Zones, which are listed in the corresponding datasets)
}
\keyword{datasets}
