% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lookups}
\alias{lookups}
\title{lookups}
\format{
A data.table with the following ??? column:
\describe{
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
}

For further details, see \url{http://}
}
\usage{
lookups
}
\description{
A list of all the geographies included in the \code{output_areas} mapping dataset (apart from the Output Areas themselves and the Workplace Zones)
}
\keyword{datasets}
