% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{neighbours}
\alias{neighbours}
\title{neighbours}
\format{
A data.table with the following two columns:
\describe{
  \item{\code{location_id}}{The code of an area}
  \item{\code{neighbour_id}}{The 1st order neighbours of an area}
}

For further details, see \url{http://}
}
\usage{
neighbours
}
\description{
This dataset contains the 1st order neighbours for all location areas
}
\keyword{datasets}
