% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{output_areas}
\alias{output_areas}
\title{Output Areas}
\format{
A data.table with the following 30 columns:
\describe{
  \item{\code{OA}}{Output Area. Theme: Census. Countries: ENSW. Total units: 232,296. Last updated: Jul-05 (Frozen). Base Layer. Direct Parent: LSOA}
  \item{\code{LSOA}}{Lower Layer Super Output Area. Theme: Census. Countries: ENSW. Total units: 42,619. Last updated: Jul-05 (Frozen). Built From OA. Direct Parent: MSOA}
  \item{\code{MSOA}}{Middle Layer Super Output Area. Theme: Census. Countries: ESW. Total units: 9,370. Last updated: Jul-05 (Frozen). Built From LSOA. Direct Parent: LAD}
  \item{\code{LAD}}{Local Authority & District. Theme: Census. Countries: ENSW. Total units: 379. Last updated: Dec-20. Built From MSOA. Direct Parent: CTY}
  \item{\code{CTY}}{County. Theme: Census. Countries: E. Total units: 93. Last updated: Jan-00. Built From LAD. Direct Parent: RGN}
  \item{\code{RGN}}{Region. Theme: Census. Countries: E. Total units: 12. Last updated: Jan-00 (Frozen). Built From CTY. Direct Parent: CTRY}
  \item{\code{CTRY}}{Country. Theme: Census. Countries: ENSW. Total units: 4. Last updated: Jan-00 (Frozen). Built From RGN. Direct Parent: 0}
  \item{\code{WPZ}}{Workplace Zone. Theme: Census. Countries: ENSW. Total units: 60,709. Last updated: Jul-05 (Frozen). Base Layer. Direct Parent: MSOA}
  \item{\code{PCS}}{Postcode Sector. Theme: Postal. Countries: ENSW. Total units: 9,538. Last updated: Nov-20. Built From OA. Direct Parent: PCD}
  \item{\code{PCD}}{Postcode District. Theme: Postal. Countries: ENSW. Total units: 2,815. Last updated: Nov-20. Built From PCS. Direct Parent: PCT}
  \item{\code{PCT}}{Post Town. Theme: Postal. Countries: ENSW. Total units: 1,430. Last updated: Nov-20. Built From PCD. Direct Parent: PCA}
  \item{\code{PCA}}{Postcode Area. Theme: Postal. Countries: ENSW. Total units: 121. Last updated: Nov-20. Built From PCT. Direct Parent: 0}
  \item{\code{PCON}}{Westminster Parliamentary Constituency. Theme: Electoral. Countries: ENSW. Total units: 650. Last updated: Jan-00. Base Layer. Direct Parent: 0}
  \item{\code{WARD}}{Electoral Ward. Theme: Electoral. Countries: ENSW. Total units: 8,875. Last updated: Dec-20. Base Layer. Direct Parent: 0}
  \item{\code{CED}}{County Electoral Division. Theme: Electoral. Countries: E. Total units: 1,719. Last updated: Jan-00 (Frozen). Base Layer. Direct Parent: 0}
  \item{\code{TTWA}}{Travel To Work Area. Theme: Urban. Countries: ENSW. Total units: 228. Last updated: Jul-05 (Frozen). Base Layer. Direct Parent: 0}
  \item{\code{MTC}}{Major Town or City. Theme: Urban. Countries: EW. Total units: 113. Last updated: Jul-05 (Frozen). Base Layer. Direct Parent: 0}
  \item{\code{BUA}}{Built-up Area. Theme: Urban. Countries: EW. Total units: 5,795. Last updated: Jul-05 (Frozen). Base Layer. Direct Parent: 0}
  \item{\code{BUAS}}{Built-up Area Sub-division. Theme: Urban. Countries: EW. Total units: 1,824. Last updated: Jul-05 (Frozen). Base Layer. Direct Parent: 0}
  \item{\code{PAR}}{Civil Parish or Unparished; Community. Theme: Social. Countries: EWS. Total units: 12,415. Last updated: Dec-20. Base Layer. Direct Parent: 0}
  \item{\code{PFN}}{Police Neighborhood. Theme: Social. Countries: ENSW. Total units: 4,814. Last updated: Nov-20. Base Layer. Direct Parent: 0}
  \item{\code{PFA}}{Police Force Area. Theme: Social. Countries: ENSW. Total units: 45. Last updated: Dec-19. Base Layer. Direct Parent: 0}
  \item{\code{FRA}}{Fire Rescue Authority. Theme: Social. Countries: ENSW. Total units: 48. Last updated: Dec-19. Base Layer. Direct Parent: 0}
  \item{\code{CSP}}{Community Safety Partnership. Theme: Social. Countries: EW. Total units: 314. Last updated: Dec-19. Base Layer. Direct Parent: 0}
  \item{\code{LPA}}{Local Planning Authority. Theme: Social. Countries: ENSW. Total units: 397. Last updated: Apr-20. Base Layer. Direct Parent: 0}
  \item{\code{RGD}}{Registration District. Theme: Social. Countries: EW. Total units: 173. Last updated: Apr-19. Base Layer. Direct Parent: 0}
  \item{\code{LRF}}{Local Resilience Forum. Theme: Social. Countries: EW. Total units: 42. Last updated: Dec-19. Base Layer. Direct Parent: 0}
  \item{\code{CCG}}{Clinical Commissioning Group. Theme: Health. Countries: ENSW. Total units: 181. Last updated: Jan-00. Built From LSOA. Direct Parent: STP}
  \item{\code{STP}}{Sustainability and Transformation Partnership. Theme: Health. Countries: E. Total units: 42. Last updated: Jan-00. Built From CCG. Direct Parent: NHSO}
  \item{\code{NHSO}}{NHS England Local Office. Theme: Health. Countries: E. Total units: 14. Last updated: Jan-00. Built From STP. Direct Parent: NHSR}
  \item{\code{NHSR}}{NHS England Region. Theme: Health. Countries: E. Total units: 7. Last updated: Jan-00. Built From NHSO. Direct Parent: 0}
  \item{\code{CIS}}{Covid Infection Survey. Theme: Health. Countries: ENSW. Total units: 133. Last updated: Jan-21. Built From OA. Direct Parent: 0}
  \item{\code{x_lon}}{}
  \item{\code{y_lat}}{}
  \item{\code{wx_lon}}{}
  \item{\code{wy_lat}}{}
  \item{\code{perimeter}}{}
  \item{\code{area}}{}
}

For further details, see \url{http://}
}
\usage{
output_areas
}
\description{
This dataset contains both a complete list of the UK \emph{Output Areas}, the smallest statistical geographic area in the UK,
with some of its characteristics, and a mapping between them and all the other Geographies contained in the package
(apart from Workplace Zones, which have a similar dedicated mapping dataset), as deployed and maintaned by \emph{ONS}.
}
\keyword{datasets}
