% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postcodes}
\alias{postcodes}
\title{A list of all the currently active postcodes in the UK (1,718,610 as of NOV-2020), their coordinates in WGS84 CRS,
and the corresponding Output Area, based on a *best-fit* approach}
\format{
A data.table with 4 columns:
\describe{
  \item{\code{PCU}}{postcode unit in 7-chars format}
  \item{\code{OA}}{output area ONS code}
  \item{\code{x_lon}}{longitude}
  \item{\code{y_lat}}{latitude}
}

For further details, see \url{http://}
}
\usage{
postcodes
}
\description{
A list of all the currently active postcodes in the UK (1,718,610 as of NOV-2020), their coordinates in WGS84 CRS,
and the corresponding Output Area, based on a *best-fit* approach
}
\keyword{datasets}
