% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{workplace_zones}
\alias{workplace_zones}
\title{Workplace Zones}
\format{
A data.table with the following six column:
\describe{
  \item{\code{WPZ}}{Workplace Zone}
  \item{\code{MSOA}}{Middle LAyer Super Output Area (NA for N.Ireland)}
  \item{\code{LAD}}{Local Authority}
  \item{\code{CTY}}{County (England only)}
  \item{\code{RGN}}{Region (England only)}
  \item{\code{CTRY}}{Country}
}

For further details, see \url{http://}
}
\usage{
workplace_zones
}
\description{
This dataset contains both a complete list of the UK \emph{Workplace Zones}, a similar statistical geographic area as OAs but based on workers commuting patterns,
with some of its characteristics, and a mapping between them and MSOA, so that there is a link to some higher level geographies.
}
\keyword{datasets}
