% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{entities}
\alias{entities}
\title{Entities}
\format{
A data.table with the following ??? column:
\describe{
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
}

For further details, see \url{http://}
}
\usage{
entities
}
\description{
An alternative view of the geographies included in the package, as listed in \code{location_types}, partitioned by countries
}
\keyword{datasets}
