% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hierarchies}
\alias{hierarchies}
\title{Hierarchies}
\format{
A data.table with the following ??? column:
\describe{
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
}

For further details, see \url{http://}
}
\usage{
hierarchies
}
\description{
A list of possible mapping between each geography and its possible parents
}
\keyword{datasets}
