% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hierarchies}
\alias{hierarchies}
\title{hierarchies}
\format{
A data.table with ??? column:
\describe{
  \item{\code{OA}}{Output Area}
}

For further details, see \url{http://}
}
\usage{
hierarchies
}
\description{
A list of all the geographies included in the \code{output_areas} mapping dataset (apart from the Output Areas themselves and the Workplace Zones)
}
\keyword{datasets}
