% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_types}
\alias{location_types}
\title{The list of the geographies included in the package}
\format{
A data.table with ??? columns:
\describe{
  \item{\code{OA}}{Output Area}
}

For further details, see \url{http://}
}
\usage{
location_types
}
\description{
A list of all the geographies included in the \code{output_areas} mapping dataset (apart from the Output Areas themselves and the Workplace Zones)
}
\keyword{datasets}
