% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{location_types}
\alias{location_types}
\title{Location Types}
\format{
A data.table with the following ??? columns:
\describe{
  \item{\code{location_type}}{An acronym for the geography, used as its id}
  \item{\code{prog_id}}{}
  \item{\code{pd_id}}{}
  \item{\code{ordering}}{}
  \item{\code{name}}{}
  \item{\code{theme}}{}
  \item{\code{countries}}{}
  \item{\code{count_ons}}{}
  \item{\code{count_pd}}{}
  \item{\code{count_pkg}}{}
  \item{\code{last_update}}{}
  \item{\code{is_frozen}}{}
  \item{\code{max_child}}{}
  \item{\code{exact_coverage}}{}
  \item{\code{min_parent}}{}
  \item{\code{pref_filter}}{}
  \item{\code{pref_map}}{}
  \item{\code{locations}}{}
  \item{\code{lookups}}{}
  \item{\code{boundaries}}{}
}

For further details, see \url{http://}
}
\usage{
location_types
}
\description{
A list of all the Geographies included in the package
(apart from the Output Areas and the Workplace Zones, which are listed in the corresponding datasets)
}
\keyword{datasets}
