% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locations}
\alias{locations}
\title{locations}
\format{
A data.table with the following ??? columns:
\describe{
  \item{\code{type}}{}
  \item{\code{location_id}}{}
  \item{\code{name}}{}
  \item{\code{x_lon}}{}
  \item{\code{y_lat}}{}
  \item{\code{wx_lon}}{}
  \item{\code{wy_lat}}{}
  \item{\code{perimeter}}{}
  \item{\code{area}}{}
}

For further details, see \url{http://}
}
\usage{
locations
}
\description{
A list of all the areas included in the \code{output_areas} mapping dataset
(apart from the Output Areas themselves and the Workplace Zones)
together with some of , like coordinates, Perimeter, and Area
}
\keyword{datasets}
