% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locations}
\alias{locations}
\title{locations}
\format{
A data.table with ??? columns:
\describe{
  \item{\code{OA}}{Output Area}
}

For further details, see \url{http://}
}
\usage{
locations
}
\description{
A list of all the areas included in the \code{output_areas} mapping dataset (apart from the Output Areas themselves and the Workplace Zones)
}
\keyword{datasets}
