% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{output_areas}
\alias{output_areas}
\title{Output Areas list plus lookups for all the remaining UK geographies}
\format{
A data.table with 30 columns:
\describe{
  \item{\code{OA}}{Output Area}
}

For further details, see \url{http://}
}
\usage{
output_areas
}
\description{
This dataset contains both a complete list of the UK \emph{Output Areas}, the smallest statistical geographic area in the UK,
with some of its characteristics, and a mapping between them and some other geographies as deployed and maintaned by \emph{ONS}.
}
\keyword{datasets}
