% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{output_areas}
\alias{output_areas}
\title{Output Areas}
\format{
A data.table with the following 30 columns:
\describe{
  \item{\code{OA}}{Output Area}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
  \item{\code{}}{}
}

For further details, see \url{http://}
}
\usage{
output_areas
}
\description{
This dataset contains both a complete list of the UK \emph{Output Areas}, the smallest statistical geographic area in the UK,
with some of its characteristics, and a mapping between them and all the other Geographies contained in the package
(apart from Workplace Zones, which have a similar dedicated mapping dataset), as deployed and maintaned by \emph{ONS}.
}
\keyword{datasets}
