% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postcodes}
\alias{postcodes}
\title{Postcodes}
\format{
A data.table with the following four columns:
\describe{
  \item{\code{PCU}}{unit postcode in 7-chars format}
  \item{\code{is_active}}{flag that indicates if the corresponding unit postcode is currently active or terminated}
  \item{\code{OA}}{output area ONS code}
  \item{\code{x_lon}}{longitude}
  \item{\code{y_lat}}{latitude}
}

For further details, see \url{http://}
}
\usage{
postcodes
}
\description{
A list of all the postcode units in the UK, as of NOV-2020 (1,718,610 active, terminated)
their coordinates in WGS84 CRS,
and the corresponding Output Area, based on a *best-fit* approach
}
\keyword{datasets}
